MueLu: Reshuffle TentativePFactory* to make diff easier
