+++ packages/muelu/src/Transfers/Smoothed-Aggregation/MueLu_TentativePFactory_kokkos_def.hpp
@@ -10,8 +10,7 @@
 #ifndef MUELU_TENTATIVEPFACTORY_KOKKOS_DEF_HPP
 #define MUELU_TENTATIVEPFACTORY_KOKKOS_DEF_HPP
 
-#include "Kokkos_UnorderedMap.hpp"
-#include "Xpetra_CrsGraphFactory.hpp"
+#include <Xpetra_CrsGraphFactory.hpp>
 
 #include "MueLu_TentativePFactory_kokkos_decl.hpp"
 
@@ -20,315 +19,18 @@
 #include "MueLu_AmalgamationFactory.hpp"
 
 #include "MueLu_MasterList.hpp"
-#include "MueLu_PerfUtils.hpp"
 #include "MueLu_Monitor.hpp"
-
-#include "Xpetra_IO.hpp"
+#include "MueLu_PerfUtils.hpp"
+#include "MueLu_Utilities.hpp"
+#include "MueLu_LocalQR.hpp"
 
 namespace MueLu {
 
-namespace {  // anonymous
+template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
+TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::TentativePFactory_kokkos() = default;
 
-template <class LocalOrdinal, class View>
-class ReduceMaxFunctor {
- public:
-  ReduceMaxFunctor(View view)
-    : view_(view) {}
-
-  KOKKOS_INLINE_FUNCTION
-  void operator()(const LocalOrdinal& i, LocalOrdinal& vmax) const {
-    if (vmax < view_(i))
-      vmax = view_(i);
-  }
-
-  KOKKOS_INLINE_FUNCTION
-  void join(LocalOrdinal& dst, const LocalOrdinal& src) const {
-    if (dst < src) {
-      dst = src;
-    }
-  }
-
-  KOKKOS_INLINE_FUNCTION
-  void init(LocalOrdinal& dst) const {
-    dst = 0;
-  }
-
- private:
-  View view_;
-};
-
-// local QR decomposition
-template <class LOType, class GOType, class SCType, class DeviceType, class NspType, class aggRowsType, class maxAggDofSizeType, class agg2RowMapLOType, class statusType, class rowsType, class rowsAuxType, class colsAuxType, class valsAuxType>
-class LocalQRDecompFunctor {
- private:
-  typedef LOType LO;
-  typedef GOType GO;
-  typedef SCType SC;
-
-  typedef typename DeviceType::execution_space execution_space;
-  typedef typename KokkosKernels::ArithTraits<SC>::val_type impl_SC;
-  typedef KokkosKernels::ArithTraits<impl_SC> impl_ATS;
-  typedef typename impl_ATS::magnitudeType Magnitude;
-
- public:
-  typedef Kokkos::View<impl_SC**, typename execution_space::scratch_memory_space, Kokkos::MemoryUnmanaged> shared_matrix;
-  typedef Kokkos::View<impl_SC*, typename execution_space::scratch_memory_space, Kokkos::MemoryUnmanaged> shared_vector;
-
- private:
-  NspType fineNS;
-  NspType coarseNS;
-  aggRowsType aggRows;
-  maxAggDofSizeType maxAggDofSize;  //< maximum number of dofs in aggregate (max size of aggregate * numDofsPerNode)
-  agg2RowMapLOType agg2RowMapLO;
-  statusType statusAtomic;
-  rowsType rows;
-  rowsAuxType rowsAux;
-  colsAuxType colsAux;
-  valsAuxType valsAux;
-  bool doQRStep;
-
- public:
-  LocalQRDecompFunctor(NspType fineNS_, NspType coarseNS_, aggRowsType aggRows_, maxAggDofSizeType maxAggDofSize_, agg2RowMapLOType agg2RowMapLO_, statusType statusAtomic_, rowsType rows_, rowsAuxType rowsAux_, colsAuxType colsAux_, valsAuxType valsAux_, bool doQRStep_)
-    : fineNS(fineNS_)
-    , coarseNS(coarseNS_)
-    , aggRows(aggRows_)
-    , maxAggDofSize(maxAggDofSize_)
-    , agg2RowMapLO(agg2RowMapLO_)
-    , statusAtomic(statusAtomic_)
-    , rows(rows_)
-    , rowsAux(rowsAux_)
-    , colsAux(colsAux_)
-    , valsAux(valsAux_)
-    , doQRStep(doQRStep_) {}
-
-  KOKKOS_INLINE_FUNCTION
-  void operator()(const typename Kokkos::TeamPolicy<execution_space>::member_type& thread, size_t& nnz) const {
-    auto agg = thread.league_rank();
-
-    // size of aggregate: number of DOFs in aggregate
-    auto aggSize = aggRows(agg + 1) - aggRows(agg);
-
-    const impl_SC one  = impl_ATS::one();
-    const impl_SC two  = one + one;
-    const impl_SC zero = impl_ATS::zero();
-    const auto zeroM   = impl_ATS::magnitude(zero);
-
-    int m = aggSize;
-    int n = fineNS.extent(1);
-
-    // calculate row offset for coarse nullspace
-    Xpetra::global_size_t offset = agg * n;
-
-    if (doQRStep) {
-      // Extract the piece of the nullspace corresponding to the aggregate
-      shared_matrix r(thread.team_shmem(), m, n);  // A (initially), R (at the end)
-      for (int j = 0; j < n; j++)
-        for (int k = 0; k < m; k++)
-          r(k, j) = fineNS(agg2RowMapLO(aggRows(agg) + k), j);
-#if 0
-          printf("A\n");
-          for (int i = 0; i < m; i++) {
-            for (int j = 0; j < n; j++)
-              printf(" %5.3lf ", r(i,j));
-            printf("\n");
-          }
-#endif
-
-      // Calculate QR decomposition (standard)
-      shared_matrix q(thread.team_shmem(), m, m);  // Q
-      if (m >= n) {
-        bool isSingular = false;
-
-        // Initialize Q^T
-        auto qt = q;
-        for (int i = 0; i < m; i++) {
-          for (int j = 0; j < m; j++)
-            qt(i, j) = zero;
-          qt(i, i) = one;
-        }
-
-        for (int k = 0; k < n; k++) {  // we ignore "n" instead of "n-1" to normalize
-          // FIXME_KOKKOS: use team
-          Magnitude s = zeroM, norm, norm_x;
-          for (int i = k + 1; i < m; i++)
-            s += pow(impl_ATS::magnitude(r(i, k)), 2);
-          norm = sqrt(pow(impl_ATS::magnitude(r(k, k)), 2) + s);
-
-          if (norm == zero) {
-            isSingular = true;
-            break;
-          }
-
-          r(k, k) -= norm * one;
-
-          norm_x = sqrt(pow(impl_ATS::magnitude(r(k, k)), 2) + s);
-          if (norm_x == zeroM) {
-            // We have a single diagonal element in the column.
-            // No reflections required. Just need to restor r(k,k).
-            r(k, k) = norm * one;
-            continue;
-          }
-
-          // FIXME_KOKKOS: use team
-          for (int i = k; i < m; i++)
-            r(i, k) /= norm_x;
-
-          // Update R(k:m,k+1:n)
-          for (int j = k + 1; j < n; j++) {
-            // FIXME_KOKKOS: use team in the loops
-            impl_SC si = zero;
-            for (int i = k; i < m; i++)
-              si += r(i, k) * r(i, j);
-            for (int i = k; i < m; i++)
-              r(i, j) -= two * si * r(i, k);
-          }
-
-          // Update Q^T (k:m,k:m)
-          for (int j = k; j < m; j++) {
-            // FIXME_KOKKOS: use team in the loops
-            impl_SC si = zero;
-            for (int i = k; i < m; i++)
-              si += r(i, k) * qt(i, j);
-            for (int i = k; i < m; i++)
-              qt(i, j) -= two * si * r(i, k);
-          }
-
-          // Fix R(k:m,k)
-          r(k, k) = norm * one;
-          for (int i = k + 1; i < m; i++)
-            r(i, k) = zero;
-        }
-
-#if 0
-            // Q = (Q^T)^T
-            for (int i = 0; i < m; i++)
-              for (int j = 0; j < i; j++) {
-                impl_SC tmp  = qt(i,j);
-                qt(i,j) = qt(j,i);
-                qt(j,i) = tmp;
-              }
-#endif
-
-        // Build coarse nullspace using the upper triangular part of R
-        for (int j = 0; j < n; j++)
-          for (int k = 0; k <= j; k++)
-            coarseNS(offset + k, j) = r(k, j);
-
-        if (isSingular) {
-          statusAtomic(1) = true;
-          return;
-        }
-
-      } else {
-        // Special handling for m < n (i.e. single node aggregates in structural mechanics)
-
-        // The local QR decomposition is not possible in the "overconstrained"
-        // case (i.e. number of columns in qr > number of rowsAux), which
-        // corresponds to #DOFs in Aggregate < n. For usual problems this
-        // is only possible for single node aggregates in structural mechanics.
-        // (Similar problems may arise in discontinuous Galerkin problems...)
-        // We bypass the QR decomposition and use an identity block in the
-        // tentative prolongator for the single node aggregate and transfer the
-        // corresponding fine level null space information 1-to-1 to the coarse
-        // level null space part.
-
-        // NOTE: The resulting tentative prolongation operator has
-        // (m*DofsPerNode-n) zero columns leading to a singular
-        // coarse level operator A.  To deal with that one has the following
-        // options:
-        // - Use the "RepairMainDiagonal" flag in the RAPFactory (default:
-        //   false) to add some identity block to the diagonal of the zero rowsAux
-        //   in the coarse level operator A, such that standard level smoothers
-        //   can be used again.
-        // - Use special (projection-based) level smoothers, which can deal
-        //   with singular matrices (very application specific)
-        // - Adapt the code below to avoid zero columns. However, we do not
-        //   support a variable number of DOFs per node in MueLu/Xpetra which
-        //   makes the implementation really hard.
-        //
-        // FIXME: do we need to check for singularity here somehow? Zero
-        // columns would be easy but linear dependency would require proper QR.
-
-        // R = extended (by adding identity rowsAux) qr
-        for (int j = 0; j < n; j++)
-          for (int k = 0; k < n; k++)
-            if (k < m)
-              coarseNS(offset + k, j) = r(k, j);
-            else
-              coarseNS(offset + k, j) = (k == j ? one : zero);
-
-        // Q = I (rectangular)
-        for (int i = 0; i < m; i++)
-          for (int j = 0; j < n; j++)
-            q(i, j) = (j == i ? one : zero);
-      }
-
-      // Process each row in the local Q factor and fill helper arrays to assemble P
-      for (int j = 0; j < m; j++) {
-        LO localRow     = agg2RowMapLO(aggRows(agg) + j);
-        size_t rowStart = rowsAux(localRow);
-        size_t lnnz     = 0;
-        for (int k = 0; k < n; k++) {
-          // skip zeros
-          if (q(j, k) != zero) {
-            colsAux(rowStart + lnnz) = offset + k;
-            valsAux(rowStart + lnnz) = q(j, k);
-            lnnz++;
-          }
-        }
-        rows(localRow + 1) = lnnz;
-        nnz += lnnz;
-      }
-
-#if 0
-          printf("R\n");
-          for (int i = 0; i < m; i++) {
-            for (int j = 0; j < n; j++)
-              printf(" %5.3lf ", coarseNS(i,j));
-            printf("\n");
-          }
-
-          printf("Q\n");
-          for (int i = 0; i < aggSize; i++) {
-            for (int j = 0; j < aggSize; j++)
-              printf(" %5.3lf ", q(i,j));
-            printf("\n");
-          }
-#endif
-    } else {
-      /////////////////////////////
-      //      "no-QR" option     //
-      /////////////////////////////
-      // Local Q factor is just the fine nullspace support over the current aggregate.
-      // Local R factor is the identity.
-      // TODO I have not implemented any special handling for aggregates that are too
-      // TODO small to locally support the nullspace, as is done in the standard QR
-      // TODO case above.
-
-      for (int j = 0; j < m; j++) {
-        LO localRow     = agg2RowMapLO(aggRows(agg) + j);
-        size_t rowStart = rowsAux(localRow);
-        size_t lnnz     = 0;
-        for (int k = 0; k < n; k++) {
-          const impl_SC qr_jk = fineNS(localRow, k);
-          // skip zeros
-          if (qr_jk != zero) {
-            colsAux(rowStart + lnnz) = offset + k;
-            valsAux(rowStart + lnnz) = qr_jk;
-            lnnz++;
-          }
-        }
-        rows(localRow + 1) = lnnz;
-        nnz += lnnz;
-      }
-
-      for (int j = 0; j < n; j++)
-        coarseNS(offset + j, j) = one;
-    }
-  }
-};
-
-}  // namespace
+template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
+TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::~TentativePFactory_kokkos() = default;
 
 template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
 RCP<const ParameterList> TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::GetValidParameterList() const {
@@ -338,7 +40,9 @@ RCP<const ParameterList> TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOr
   SET_VALID_ENTRY("tentative: calculate qr");
   SET_VALID_ENTRY("tentative: build coarse coordinates");
   SET_VALID_ENTRY("sa: keep tentative prolongator");
+  // SET_VALID_ENTRY("tentative: constant column sums");
 #undef SET_VALID_ENTRY
+  validParamList->set<std::string>("Nullspace name", "Nullspace", "Name for the input nullspace");
 
   validParamList->set<RCP<const FactoryBase>>("A", Teuchos::null, "Generating factory of the matrix A");
   validParamList->set<RCP<const FactoryBase>>("Aggregates", Teuchos::null, "Generating factory of the aggregates");
@@ -389,13 +93,24 @@ void TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP
   FactoryMonitor m(*this, "Build", coarseLevel);
 
   typedef typename Teuchos::ScalarTraits<Scalar>::coordinateType coordinate_type;
+  typedef Xpetra::MultiVector<coordinate_type, LO, GO, NO> RealValuedMultiVector;
   typedef Xpetra::MultiVectorFactory<coordinate_type, LO, GO, NO> RealValuedMultiVectorFactory;
+
   const ParameterList& pL = GetParameterList();
   std::string nspName     = "Nullspace";
   if (pL.isParameter("Nullspace name")) nspName = pL.get<std::string>("Nullspace name");
 
-  auto A             = Get<RCP<Matrix>>(fineLevel, "A");
-  auto aggregates    = Get<RCP<Aggregates>>(fineLevel, "Aggregates");
+  RCP<Matrix> Ptentative;
+  auto A          = Get<RCP<Matrix>>(fineLevel, "A");
+  auto aggregates = Get<RCP<Aggregates>>(fineLevel, "Aggregates");
+  // No coarse DoFs so we need to bail by setting Ptentattive to null and returning
+  //  This level will ultimately be removed in MueLu_Hierarchy_defs.h via a resize()
+  if (aggregates->GetNumGlobalAggregatesComputeIfNeeded() == 0) {
+    Ptentative = Teuchos::null;
+    Set(coarseLevel, "P", Ptentative);
+    return;
+  }
+
   auto amalgInfo     = Get<RCP<AmalgamationInfo>>(fineLevel, "UnAmalgamationInfo");
   auto fineNullspace = Get<RCP<MultiVector>>(fineLevel, nspName);
   auto coarseMap     = Get<RCP<const Map>>(fineLevel, "CoarseMap");
@@ -404,14 +119,16 @@ void TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP
     fineCoords = Get<RCP<RealValuedMultiVector>>(fineLevel, "Coordinates");
   }
 
-  RCP<Matrix> Ptentative;
-  // No coarse DoFs so we need to bail by setting Ptentattive to null and returning
-  // This level will ultimately be removed in MueLu_Hierarchy_defs.h via a resize()
-  if (aggregates->GetNumGlobalAggregatesComputeIfNeeded() == 0) {
-    Ptentative = Teuchos::null;
-    Set(coarseLevel, "P", Ptentative);
-    return;
+  // FIXME: We should remove the NodeComm on levels past the threshold
+  if (fineLevel.IsAvailable("Node Comm")) {
+    RCP<const Teuchos::Comm<int>> nodeComm = Get<RCP<const Teuchos::Comm<int>>>(fineLevel, "Node Comm");
+    Set<RCP<const Teuchos::Comm<int>>>(coarseLevel, "Node Comm", nodeComm);
   }
+
+  // NOTE:  We check DomainMap here rather than RowMap because those are different for BlockCrs matrices
+  TEUCHOS_TEST_FOR_EXCEPTION(A->getDomainMap()->getLocalNumElements() != fineNullspace->getMap()->getLocalNumElements(),
+                             Exceptions::RuntimeError, "MueLu::TentativePFactory::MakeTentative: Size mismatch between A and Nullspace");
+
   RCP<MultiVector> coarseNullspace;
   RCP<RealValuedMultiVector> coarseCoords;
 
@@ -461,7 +178,7 @@ void TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP
       typename AppendTrait<decltype(fineCoordsView), Kokkos::RandomAccess>::type fineCoordsRandomView = fineCoordsView;
       for (size_t j = 0; j < dim; j++) {
         Kokkos::parallel_for(
-            "MueLu::TentativeP::BuildCoords", Kokkos::RangePolicy<local_ordinal_type, execution_space>(0, numAggs),
+            "MueLu::TentativeP::BuildCoords", Kokkos::RangePolicy<LocalOrdinal, execution_space>(0, numAggs),
             KOKKOS_LAMBDA(const LO i) {
               // A row in this graph represents all node ids in the aggregate
               // Therefore, averaging is very easy
@@ -502,13 +219,6 @@ void TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP
   if (bTransferCoordinates_) {
     Set(coarseLevel, "Coordinates", coarseCoords);
   }
-
-  // FIXME: We should remove the NodeComm on levels past the threshold
-  if (fineLevel.IsAvailable("Node Comm")) {
-    RCP<const Teuchos::Comm<int>> nodeComm = Get<RCP<const Teuchos::Comm<int>>>(fineLevel, "Node Comm");
-    Set<RCP<const Teuchos::Comm<int>>>(coarseLevel, "Node Comm", nodeComm);
-  }
-
   Set(coarseLevel, "Nullspace", coarseNullspace);
   Set(coarseLevel, "P", Ptentative);
 
@@ -539,7 +249,8 @@ void TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
   typedef KokkosKernels::ArithTraits<SC> ATS;
   using impl_SC      = typename ATS::val_type;
   using impl_ATS     = KokkosKernels::ArithTraits<impl_SC>;
-  const impl_SC zero = impl_ATS::zero(), one = impl_ATS::one();
+  const impl_SC zero = impl_ATS::zero();
+  const impl_SC one  = impl_ATS::one();
 
   const LO INVALID = Teuchos::OrdinalTraits<LO>::invalid();
 
@@ -557,7 +268,7 @@ void TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
   bool goodMap;
   {
     SubFactoryMonitor m2(*this, "Check good map", coarseLevel);
-    goodMap = isGoodMap(*rowMap, *colMap);
+    goodMap = Utilities::MapsAreNested(*rowMap, *colMap);
   }
   // FIXME_KOKKOS: need to proofread later code for bad maps
   TEUCHOS_TEST_FOR_EXCEPTION(!goodMap, Exceptions::RuntimeError,
@@ -571,7 +282,10 @@ void TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
   // the data of the AmalgamationInfo container class
 
   // Extract information for unamalgamation
-  LO fullBlockSize, blockID, stridingOffset, stridedBlockSize;
+  LO fullBlockSize;
+  LO blockID;
+  LO stridingOffset;
+  LO stridedBlockSize;
   GO indexBase;
   amalgInfo->GetStridingInformation(fullBlockSize, blockID, stridingOffset, stridedBlockSize, indexBase);
   GO globalOffset = amalgInfo->GlobalOffset();
@@ -630,7 +344,7 @@ void TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
   // Find maximum dof size for aggregates
   // Later used to reserve enough scratch space for local QR decompositions
   LO maxAggSize = 0;
-  ReduceMaxFunctor<LO, decltype(aggDofSizes)> reduceMax(aggDofSizes);
+  LocalQR::ReduceMaxFunctor<LO, decltype(aggDofSizes)> reduceMax(aggDofSizes);
   Kokkos::parallel_reduce("MueLu:TentativePF:Build:max_agg_size", range_type(0, aggDofSizes.extent(0)), reduceMax, maxAggSize);
 
   // parallel_scan (exclusive)
@@ -838,10 +552,10 @@ void TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
       // Each team handles a slice of the data associated with one aggregate
       // and performs a local QR decomposition
       Kokkos::TeamPolicy<execution_space> policy(numAggregates, 1);  // numAggregates teams a 1 thread
-      LocalQRDecompFunctor<LocalOrdinal, GlobalOrdinal, Scalar, DeviceType, decltype(fineNSRandom),
-                           decltype(aggDofSizes /*aggregate sizes in dofs*/), decltype(maxAggSize), decltype(agg2RowMapLO),
-                           decltype(statusAtomic), decltype(rows), decltype(rowsAux), decltype(colsAux),
-                           decltype(valsAux)>
+      LocalQR::LocalQRDecompFunctor<LocalOrdinal, GlobalOrdinal, Scalar, DeviceType, decltype(fineNSRandom),
+                                    decltype(aggDofSizes /*aggregate sizes in dofs*/), decltype(maxAggSize), decltype(agg2RowMapLO),
+                                    decltype(statusAtomic), decltype(rows), decltype(rowsAux), decltype(colsAux),
+                                    decltype(valsAux)>
           localQRFunctor(fineNSRandom, coarseNS, aggDofSizes, maxAggSize, agg2RowMapLO, statusAtomic,
                          rows, rowsAux, colsAux, valsAux, doQRStep);
       if (doQRStep) {
@@ -1013,7 +727,7 @@ void TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
   // Aggregates map is based on the amalgamated column map
   // We can skip global-to-local conversion if LIDs in row map are
   // same as LIDs in column map
-  bool goodMap = MueLu::Utilities<SC, LO, GO, NO>::MapsAreNested(*rowMap, *colMap);
+  bool goodMap = Utilities::MapsAreNested(*rowMap, *colMap);
   TEUCHOS_TEST_FOR_EXCEPTION(!goodMap, Exceptions::RuntimeError,
                              "MueLu: TentativePFactory_kokkos: for now works only with good maps "
                              "(i.e. \"matching\" row and column maps)");
@@ -1025,7 +739,10 @@ void TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
   // the data of the AmalgamationInfo container class
 
   // Extract information for unamalgamation
-  LO fullBlockSize, blockID, stridingOffset, stridedBlockSize;
+  LO fullBlockSize;
+  LO blockID;
+  LO stridingOffset;
+  LO stridedBlockSize;
   GO indexBase;
   amalgInfo->GetStridingInformation(fullBlockSize, blockID, stridingOffset, stridedBlockSize, indexBase);
   // GO globalOffset = amalgInfo->GlobalOffset();
@@ -1055,7 +772,7 @@ void TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
   // Find maximum dof size for aggregates
   // Later used to reserve enough scratch space for local QR decompositions
   LO maxAggSize = 0;
-  ReduceMaxFunctor<LO, decltype(aggDofSizes)> reduceMax(aggDofSizes);
+  LocalQR::ReduceMaxFunctor<LO, decltype(aggDofSizes)> reduceMax(aggDofSizes);
   Kokkos::parallel_reduce("MueLu:TentativePF:Build:max_agg_size", range_type(0, aggDofSizes.extent(0)), reduceMax, maxAggSize);
 
   // parallel_scan (exclusive)
@@ -1261,29 +978,6 @@ void TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
                   RCP<const Map> /* coarseMap */, RCP<Matrix>& /* Ptentative */,
                   RCP<MultiVector>& /* coarseNullspace */) const {
   throw Exceptions::RuntimeError("MueLu: Construction of coupled tentative P is not implemented");
-}
-
-template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-bool TentativePFactory_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
-    isGoodMap(const Map& rowMap, const Map& colMap) const {
-  auto rowLocalMap = rowMap.getLocalMap();
-  auto colLocalMap = colMap.getLocalMap();
-
-  const size_t numRows = rowLocalMap.getLocalNumElements();
-  const size_t numCols = colLocalMap.getLocalNumElements();
-
-  if (numCols < numRows)
-    return false;
-
-  size_t numDiff = 0;
-  Kokkos::parallel_reduce(
-      "MueLu:TentativePF:isGoodMap", range_type(0, numRows),
-      KOKKOS_LAMBDA(const LO i, size_t& diff) {
-        diff += (rowLocalMap.getGlobalElement(i) != colLocalMap.getGlobalElement(i));
-      },
-      numDiff);
-
-  return (numDiff == 0);
 }
 
 }  // namespace MueLu
